pyproject.toml
[project]
name = "smart-gallery-guide-bot"
version = "0.1.0"
description = "Smart gallery guide bot in Telegram format"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ivan Sidorov"}
]
keywords = ["multimodal", "art", "vllm", "llm", "rag", "telegram", "bot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-telegram-bot>=22.0",
    "aiogram>=3.19.0",

    "chromadb>=1.0.0",
    "faiss-cpu>=1.8.0",

    "torch>=2.2.2",
    "transformers>=4.51.0",
    "sentence-transformers>=4.0.0",
    "Pillow>=11.2.1",
    # "vllm>=0.8.0",

    "openai>=1.70.0",

    "python-dotenv>=1.0.0",
    "numpy>=2.2.4",
    "aiofiles>=23.2.1",
    "pydantic>=2.11.0",
]

[project.scripts]
smart-guide-bot = "bot.bot:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent", "bot", "config", "database", "models"]
